Minor fixes.
TEST THE CODE AFTER MAKING CHANGES! If the code doesn't run after the
changes you made, comment out those changes and give reasons or errors in
the commit entry.

AnimatedSprite DOES NOT have set_animation_speed. It's in SpriteFrames.
SpriteFrames is a frame library class for AnimatedSprite. Accessed in
broadly a similar way to a SampleLibrary.

Fixing the code to use the SpriteFrames resource properly breaks the
animation.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>