Replaced the test platform with another test platform.
The difference is the new one is a RigidBody2D, so should provide for a
surface when implementing gravity, floors and so on. Also need to look at
the 2D platformer demo to see how it does platforms, especially moving
ones.

Also: keep collision shapes on in the editor while making this. And some
minor changes to vector stuff.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>