Rings are in!
Got rings up and running. Nothing fancy, they just vanish on contact with
the player (ideally, it should be the ring making the check, but it does
not seem to work unless it's in Sonic's script). At this point it should
then be easy enough to add the functionality needed.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>