Checkpoints!
Checkpoints are added. The way it should work is to reset the player's
position to the checkpoint's position. It does that by setting the
checkpoint_pos vector within the player's properties to the position of
the checkpoint.

The test checkpoint works, and for testing purposes the jump action will
reset the player to the checkpoint's position. Will require further
testing though!

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>