Backgrounds and other things.
Shoved the background there is at the moment into a ParallaxBackground and
ParallaxLayer. Won't make any difference for the background as is, but
seeing how they work for use with more involved (multiple image)
backdrops in future.

Added a stopwatch symbol for the timer, commented a bit more code, added
setters and getters for the score as well.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>