Godot 3 time!
Forked the original source tree and ported it over to Godot 3. It broadly
appears to work now as the original 2.x version has. Still a lot to do with
it though. Input might need some reworking (more efficient, and to work
better with the way 3 does things) for a start. Ensure "physics" are
applied to the player character; at the moment he floats. And more.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>