Rings changed!
Rings are now Area2D nodes. Stops the horde of error messages at least!

Next change is for the player - going to be using the DCC 2D platformer
approach and making the player a RigidBody2D. Might be a bit more involved
with movement code and the like, but should offer sufficient gains. Shall
see!

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>