Lives system, HUD changes and more?
The lives system appears to be working as it should, via the getter and
setter functions. All you need to do is to increase/decrease the lives
variable (remembering self. if needed!) and it'll automatically do the
rest.

HUD has seen some minor changes and stuff behind-the-scenes. Lives count
will flash if the player is on their last life. Dunno if it'd be too
distracting, but for now it stays. Minor node re-org to take that into
account.


Added some debugging keys. Probably should add some more at some point
soon.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>