Lives on 100x rings! Plus...
Well, the every-100-rings-gets-you-an-extra-life mechanism now works as it
should. Probably not very optimally, but it works so there you go!

A game-space singleton has been added. Stuff is going to be shoved over to
it as necessary to keep crap out of the scene/node scripts as much as is
possible.

Also, some of the filenames have been changed for consistency's sake.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>