Stabbing the movement code!
Trying to get the movement code into some kind of line. At the moment it
detects the floor, but it isn't really that good. The original movement
code was not written in a way to take delta into account, which has
resulted in some oddities. So it needs total re-writing. The demo suite
platformer demos certainly point the way to go, but as they stand they'd
also be more Mario-like than I want with this project.

Anyway, this code sort-of works as-is, but really shouldn't be used for
anything except as a "how not to" guide! It needs re-working from the
ground up.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>