RigidBody2D misadventures ahoy!
Trying RigidBody2D to see how things go. So far, not terribly well! I need
to get in and actually understand how things are working, and what they are
doing and why. Some of it is due to scaling I'd guess. Going to have to try
and re-adjust it. Needs to be adjusted to take into account player's speed
anyway (such as animations, maximum speeds and so on).

TO FIX: Acceleration, deceleration, jumping, animations.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>