Minor new stuff.
Should stop trying to filter new images/textures on import/re-import now.

Player jumping altered a little, not that it makes much of a difference,
need to check mass and other settings on the character (NOT player_generic
though!).

Dunno why, but made the temporary start screen a tiny bit less ugly.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>