Checkpoints are working again!
Starting to re-implement the debugging commands and noticed that the
checkpoints weren't working as they should. They should do now.

Got a very strange issue where the HUD isn't being hidden if the player
jumps off the ledge, keeps the movement key held down and then triggers
the death debug key. Not sure how much of an issue that will be later on,
shall have to see.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>