Inconsistency found?
As per https://github.com/godotengine/godot/issues/15218

Maybe. The queue_free line is somewhere different (and not the last line
of the block it's in either). Now, this indicates that perhaps the Windows
behaviour is correct and the Linux behaviour incorrect, but they're still
not consistent (should be the same on both platforms, should it not?).

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>