The commenting! Plus other things.
Just stuck some placeholder music for the "title screen". Not needed, but
it's there. Fixed up some commenting for the source as well, more work is
needed but it's just about functional. music_player can now play from any
point in a music file.

Lots more still needs to be done - like clamping to ramps for one, making
the player character not floaty for another, and so on. These things are
next!

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>