Platforms! Structure! etc.
Well, deleted the temporary platform and replaced it with the tileset. It
has some collision issues though, need to take a look and fix up as
needed.

Added the Structure.txt so that if anyone else looks through the code they
have a guide as to how it's laid out (not that it should be that hard to
figure out, but!).

Removed some more duplicated code. Hopefully it makes things a tiny bit
easier to read?

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>