Path of the commitments?
Look at how Path2D and its associated nodes work and use them for things
like loops. May also work for slopes, don't know yet but might be worth a
go. Whether it'd actually work for a "proper" Sonic game is unknown (and,
honestly, doubtful) but it's of use to take a look anyway.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>